+++ R/extract_abs_sheets.R
@@ -1,19 +1,18 @@
-
 #' Extract data sheets from an ABS timeseries workbook saved locally as an
 #' .xls file.
 #'
-#' Note that this function will not tidy the data for you. Use `read_abs_local()`
-#' to import and tidy data from local ABS time series spreadsheets or `read_abs()`
-#' to download, import and tidy ABS time series.
+#' Note that this function will not tidy the data for you.
+#' Use `read_abs_local()`to import and tidy data from local ABS time series
+#' spreadsheets or `read_abs()` to download, import and tidy ABS time series.
 #'
-#' @param filename Filename for an ABS time series spreadsheet (as string), such as
-#' "6202002.xls".
+#' @param filename Filename for an ABS time series spreadsheet (as string),
+#' such as "6202002.xls".
 #'
 #' @param table_title String giving the full title of the ABS table, such as
 #' "Table 1. Employed persons, Australia"
 #'
 #' @param path Local directory in which an ABS time series is stored. Default is
-#' "data/ABS"
+#' `Sys.getenv("R_READABS_PATH", unset = tempdir())`.
 #'
 #' @importFrom readxl excel_sheets read_xls
 #' @importFrom tibble tibble
@@ -23,9 +22,13 @@
 #'
 #' @export
 
-extract_abs_sheets <- function(filename, table_title = NULL, path = "data/ABS"){
+extract_abs_sheets <- function(filename,
+                               table_title = NULL,
+                               path = Sys.getenv("R_READABS_PATH",
+                                                 unset = tempdir())) {
 
-  .=sheet=NULL
+  # Satisfy CRAN
+  . = sheet = NULL
 
   filename <- paste0(path, "/", filename)
 
@@ -33,8 +36,9 @@ extract_abs_sheets <- function(filename, table_title = NULL, path = "data/ABS"){
     tibble::tibble(sheet = .) %>%
     dplyr::filter(!sheet %in% c("Index", "Inquiries"))
 
-  if(nrow(sheets) < 1){
-    stop(sprintf("The Excel workbook %s appears to have no data sheets.", filename))
+  if (nrow(sheets) < 1) {
+    stop(sprintf("The Excel workbook %s appears to have no data sheets.",
+                 filename))
   }
 
   data_sheets <- purrr::map(.x = sheets$sheet,